4_Bigquery_Data_Warehouse/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
fsspec
gcsfs
psycopg2-binary
sqlalchemy
google-cloud-secret-manager
pandas-gbq